feat(rule-tester): Extract result to component
